Implement feature converter for a decoder.
A decoder-only architecture can be used in various ways. This CL implements the
most basic format where the masking pattern is fully causal.

The decoder only architecture can be used as "targets" only dataset (e.g.,
standard language modeling objective) or a dataset with a notion of "inputs"
and "targets" (e.g., machine translation).

Currently in the mesh_transformer implementation, the concatenation of "inputs"
and "targets" happens in the mtf.transformer.utils._dynamic_text2self. This CL
implements this functionality with tf. The new implementation is significantly
simpler because we pack the dataset after the concatenation.

PiperOrigin-RevId: 343788232